[SofaExporter] Move all files to be compliant with recommanded directory structure
The flat structure in our plugins was problematic because it was
favoring the leaking of includes out of their packages.

By putting the sources in the
src/ModuleName/*.h and exporting the src as the include path in the CMAKeLists.txt
allows to write
#include <ModuleName/thisfile.h> which is clearer.
