+++ modules/SofaExporter/src/SofaExporter/Exporter.cpp
