+++ modules/SofaExporter/src/SofaExporter/OBJExporter.cpp
