+++ modules/SofaExporter/src/SofaExporter/STLExporter.cpp
