+++ modules/SofaExporter/src/SofaExporter/WriteTopology.cpp
